requirements.txt
meshcore>=2.1.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
PyYAML>=6.0
packaging>=23.0
